[Refactor] BoardResponse 일급 컬렉션 생성
